ComplianceCheckSet#update_status: Return `true` if no update happened
Return `true` from this method by default. If the `#update` call fails,
then it will return `false`, but in all other cases, we should return
`true`. That way, when the
`Api::V1::ComplianceCheckSetsController#validated` endpoint is called,
it will respond with the `ComplianceCheckSet` object instead of:

    {"status":"error","messages":[]}

with no error messages.

Refs #4757