Referential: Move `create_schema` to an `after_commit`
The goal being to separate the creation of the database schema from the
creation of the `Referential` object. Previously these were created in
the same database transaction, but moving the schema to an
`after_commit` puts it in a separate transaction.

This change tries to mitigate a problem where two `Referential`s,
created at the same time with the same data, can be saved, even though
identical `Referential`s should not be allowed. By moving the schema
creation to a separate transaction, we decrease the time it takes to
create a `Referential`, increasing the odds of not creating two
identical `Referential`s due to near-simultaneous writes.

Refs #5024